+++ lib/src/translation_downloader/json/json_api_translation_downloader.dart
@@ -2,12 +2,9 @@ import 'package:http/http.dart' as http;
 
 import 'json_translation_downloader.dart';
 
-const _baseUrl = 'https://dashboard.impaktfull.com';
 const _baseUrlPath = '/api/apps/translations';
 
 class JsonApiTranslationDownloader extends JsonTranslationDownloader {
-  String get host => params.host ?? _baseUrl;
-
   const JsonApiTranslationDownloader(
     super.params,
   );
@@ -16,17 +13,24 @@ class JsonApiTranslationDownloader extends JsonTranslationDownloader {
   Future<String> downloadJson(String language) async {
     final headers = <String, String>{
       'accept': 'application/json',
-      'Authorization': 'Bearer ${params.apiKey}',
+      'X-IF-TRANSLATION-APP-API-KEY': params.apiKey,
     };
-    final url = '$host$_baseUrlPath/${params.appUuid}/translations/$language';
-    final uri = Uri.parse(url);
+    final uri = Uri(
+      scheme: 'https',
+      host: params.host.replaceFirst('https://', ''),
+      path: _baseUrlPath,
+      queryParameters: {
+        'translationAppUuid': params.appUuid,
+        'locales': language,
+      },
+    );
     final response = await http.get(
       uri,
       headers: headers,
     );
     if (response.statusCode != 200) {
       throw Exception(
-        '\n\nFailed to get $url with status code ${response.statusCode}\n',
+        '\n\nFailed to get ${uri.toString()} with status code ${response.statusCode}\n',
       );
     }
     return response.body;